Bugfixing1 (#41)
* fix bugs

* fix bugs

* remove useless code

* fix readme table format

* refine readme

* reformat output

* refine progress indicator

* refine progress indicator

* update gif, more clear instruction for cloudshell, fix no workspace bug

* refine doc words, remove useless reference

* fix typo

* remove useless export

* continue refine readme

* log error output

+++ src/cloudShellRunner.ts
@@ -17,6 +17,10 @@ const tempFile = path.join(ost.tmpdir(), 'cloudshell' + vscode.env.sessionId + '
 
 export class CloudShellRunner extends BaseRunner {
 
+    constructor(outputChannel: vscode.OutputChannel) {
+        super(outputChannel);
+    }
+
     protected runPlaybookInternal(playbook: string): void {
         const installedExtension: any[] = vscode.extensions.all;
 
@@ -56,25 +60,29 @@ export class CloudShellRunner extends BaseRunner {
 
                     openCloudConsole(accountApi, OSes.Linux, [playbook], this._outputChannel, tempFile).then(terminal => {
                         var count = 30;
-                        const interval = setInterval(function () {
-                            count--;
-                            if (count > 0) {
-                                if (fsExtra.existsSync(tempFile)) {
-                                    fsExtra.removeSync(tempFile);
-                                    if (utilities.isTelemetryEnabled()) {
-                                        terminal.sendText('export ' + Constants.UserAgentName + '=' + utilities.getUserAgent());
+                        if (terminal) {
+                            var _localthis = this;
+                            var interval = setInterval(function () {
+                                count--;
+                                if (count > 0) {
+                                    if (fsExtra.existsSync(tempFile)) {
+                                        fsExtra.removeSync(tempFile);
+                                        if (utilities.isTelemetryEnabled()) {
+                                            terminal.sendText('export ' + Constants.UserAgentName + '=' + utilities.getUserAgent());
+                                        }
+                                        terminal.sendText('ansible-playbook ' + path.basename(playbook));
+                                        terminal.show();
+
+                                        count = 0;
                                     }
-                                    terminal.sendText('ansible-playbook ' + path.basename(playbook));
-                                    terminal.show();
-
-                                    count = 0;
+                                } else {
+                                    _localthis.stop(interval);
                                 }
-                            } else {
-                                this.stop(interval);
-                            }
-                        }, 500);
+                            }, 500);
+                        } else {
+                            this._outputChannel.appendLine('\nConnecting to terminal failed, please retry.');
+                        }
                     });
-
 
                 } else if (response === cancelItem) {
                     opn('https://docs.microsoft.com/en-us/azure/cloud-shell/pricing');
